Feat: Updating GameContracts to include SessionId for PDA

+++ programs/vortex-contracts/src/controllers/game_stake.rs
@@ -1,12 +1,12 @@
 
 
 pub fn calculate_winner_and_bettor_rewards(
-    total_money_staked_by_user: u64,
-    total_money_staked_on_player: u64,
-    total_pot: u64,
-) -> u64{
+    total_money_staked_by_user: f64,
+    total_money_staked_on_player: f64,
+    total_pot: f64,
+) -> f64{
 
-    let user_ratio_staked =( (total_money_staked_by_user) / (total_money_staked_on_player)) as u64;
+    let user_ratio_staked =( (total_money_staked_by_user) / (total_money_staked_on_player)) as f64;
 
     user_ratio_staked * total_pot
